pyproject.toml
[project]
name = "django_comments"
version = "0.1.0"
dependencies = [
    "Django==4.2",
    "django-filter>=23.5",
    "django-dotenv>=1.4",
    "djangorestframework>=3.14",
    "drf-recaptcha>=3.0",
    "bleach>=6.1",
    "loguru>=0.7",
    "psycopg2-binary>=2.9.9",
]


[project.optional-dependencies]
dev = [
    "black>=23.1",
    "isort>=5.12",
    "pudb>=2023.1",
    "ruff>=0.0.261",
    "types-requests>=2.31",
]


[tool.pip-tools]
generate_hashes = true


[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".mypy_cache",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "__pycache_",
    "_build",
    "buck-out",
    "build",
    "dist",
    "venv",
    "Makefile",
    "migrations",
]


# Same as Black.
line-length = 79

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py311"


[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10


[tool.black]
target-version = ['py311']
line-length = 79
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79
skip = '.venv,venv,env'
src_paths = ["src"]


[tool.coverage.run]
omit = [
    "*/venv/",
    "*/conftest.py",
    "*/migrations/",
]


[tool.pyright]
exclude = ["**/__pycache__", "**/venv/*", "**/migrations/*"]
typeCheckingMode = "off"  # off, basic, strict
useLibraryCodeForTypes = true
